5 27 21 2am push
all things present here for the vid recording.

NOTES:
There's some jank with the dronedeath event, but im p sure its b/c the drone is just kinda janky atm. (ive taken some measures in fmod to rectify this, but the way the drone is set up u can just keep killing it after it's dead. [Remember that it takes only 5 bullets to kill the drone, and any more will kill it again])
Music event is still scuffed, though not enough for people to notice in the vid.
theres an error that pops up saying that the music event cound not be found in the bank, but this doesnt seem to do anything. Music event still functions as intended (previously mentioned scuff notwithstanding)